Refactor: move ErrorDialog to uiblocks folder
